requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
pdfplumber>=0.10.3
python-docx>=1.1.0
pymupdf>=1.23.8
pillow>=10.0.0
textract>=1.6.5
rapidfuzz>=3.1.1
pytesseract>=0.3.10
openai>=1.30.5
python-dotenv>=1.0.1
